refactor: 비즈니스 로직 커스텀 훅으로 분리 및 내보내기 방식 변경으로 인한 버그 수정
