+++ client/src/app/(protected)/setting/page.tsx
@@ -1,22 +1,25 @@
 'use client';
 
-import { useRouter } from 'next/navigation';
 import { TopBar } from '@/app/components/SettingTopbar/TopBar';
 import { UserInfo } from '@/app/components/UserInfo/UserInfo';
 import { Option } from '@/app/components/Option/Option';
 import BottomSheet from '@/app/components/BottomSheet/BottomSheet';
-import { useState } from 'react';
 import BottomSheetContent from '@/app/components/BottomSheet/BottomSheetContent';
-import { useAuthStore } from '@/app/stores/authStore';
 import { SettingPageContainer, Content, Divider } from './styles';
+import { useSetting } from '@/app/hooks/useSetting';
 
 export const SettingPage = () => {
-  const router = useRouter();
-  const [isOpen, setIsOpen] = useState(false);
-  const { nickname, email, logout } = useAuthStore();
-  const handleLogout = () => {
-    setIsOpen(true);
-  };
+  const {
+    isOpen,
+    nickname,
+    email,
+    handleLogout,
+    handleClose,
+    handleConfirmLogout,
+    handleNicknameChange,
+    handlePasswordChange,
+  } = useSetting();
+
   return (
     <SettingPageContainer>
       <TopBar />
@@ -32,12 +35,12 @@ export const SettingPage = () => {
         <Option
           iconSrc="/images/user.png"
           label="닉네임 변경"
-          onClick={() => router.push('/nickname')}
+          onClick={handleNicknameChange}
         />
         <Option
           iconSrc="/images/lock.png"
           label="비밀번호 변경"
-          onClick={() => router.push('/password')}
+          onClick={handlePasswordChange}
         />
         <Option
           iconSrc="/images/unlock.png"
@@ -51,12 +54,9 @@ export const SettingPage = () => {
               leftIcon="/images/cancel.png"
               leftButtonText="취소"
               rightButtonText="로그아웃"
-              onClose={() => setIsOpen(false)}
-              onLeftClick={() => setIsOpen(false)}
-              onRightClick={() => {
-                logout();
-                router.push('/');
-              }}
+              onClose={handleClose}
+              onLeftClick={handleClose}
+              onRightClick={handleConfirmLogout}
             />
           </BottomSheet>
         )}
